lsp: simplify yamlls settings
K8s schemas are still a real PITA to setup and just won't work properly.

Options described in https://www.arthurkoziel.com/json-schemas-in-neovim/
have to much manual interaction and yaml-companion seems dead.

A promising option is https://github.com/diogo464/kubernetes.nvim, but
in this setup it is not creating the schema.json file for some reason.

For now disable yamlls validation and only enable completion for
standard k8s resources (hardcoded to v1.22 in yamlls).